requirements.txt
aiogram>=3.0.0
aiohttp
beautifulsoup4
aiosqlite
matplotlib
python-dotenv
fake-useragent